feat: add new package for react data utilities
