+++ packages/fast-data-utilities-react/.prettierignore
@@ -1,0 +1,2 @@
+coverage/*
+dist/*
