+++ packages/fast-data-utilities-react/package.json
@@ -1,0 +1,91 @@
+{
+    "name": "@microsoft/fast-data-utilities-react",
+    "description": "FAST data utilities for React",
+    "sideEffects": false,
+    "version": "1.0.0",
+    "author": {
+      "name": "Microsoft",
+      "url": "https://discord.gg/FcSNfg4"
+    },
+    "license": "MIT",
+    "repository": {
+      "type": "git",
+      "url": "git+https://github.com/Microsoft/fast-dna.git"
+    },
+    "bugs": {
+      "url": "https://github.com/Microsoft/fast-dna/issues/new/choose"
+    },
+    "main": "dist/index.js",
+    "types": "dist/index.d.ts",
+    "scripts": {
+      "build": "tsc -p ./tsconfig.json",
+      "clean:dist": "node ../../build/clean.js dist",
+      "prepare": "npm run clean:dist && npm run build",
+      "prettier": "prettier --config ../../.prettierrc --write '**/*.{ts,tsx}'",
+      "prettier:diff": "prettier --config ../../.prettierrc '**/*.{ts,tsx}' --list-different",
+      "test": "npm run tslint && npm run unit-tests",
+      "tslint": "tslint -c ./tslint.json '**/*.ts{,x}'",
+      "tslint:fix": "tslint -c ./tslint.json --fix '**/*.ts{,x}'",
+      "unit-tests": "jest --maxWorkers=4",
+      "watch": "npm run build -- -w --preserveWatchOutput"
+    },
+    "jest": {
+        "collectCoverage": true,
+        "coverageThreshold": {
+          "global": {
+            "statements": 80,
+            "branches": 80,
+            "functions": 80,
+            "lines": 80
+          }
+        },
+        "coveragePathIgnorePatterns": [
+          "/.tmp/*"
+        ],
+        "testURL": "http://localhost",
+        "transform": {
+          "^.+\\.tsx?$": "ts-jest",
+          "^.+\\.jsx?$": "babel-jest"
+        },
+        "transformIgnorePatterns": [
+          "!<rootDir>/node_modules/lodash-es"
+        ],
+        "testRegex": "(\\.|/)(test|spec)\\.(jsx?|tsx?)$",
+        "moduleFileExtensions": [
+          "ts",
+          "tsx",
+          "js",
+          "jsx"
+        ]
+      },
+    "dependencies": {
+      "tv4": "^1.3.0",
+      "lodash-es": "^4.0.0"
+    },
+    "devDependencies": {
+      "@microsoft/fast-tslint-rules": "^3.0.0",
+      "@babel/core": "^7.0.0",
+      "@babel/preset-env": "^7.1.0",
+      "@babel/preset-react": "^7.0.0",
+      "@types/jest": "^23.3.5",
+      "babel-jest": "^23.6.0",
+      "babel-core": "^7.0.0-bridge.0",
+      "jest": "^23.6.0",
+      "ts-jest": "^23.10.4",
+      "@types/lodash-es": "^4.17.0",
+      "@types/node": "^9.6.7",
+      "@types/react": "^16.3.6",
+      "lodash-es": "^4.0.0",
+      "prettier": "1.14.3",
+      "react": "^16.3.0",
+      "rimraf": "^2.6.2",
+      "tslint": "^5.9.1",
+      "tslint-config-prettier": "^1.15.0",
+      "tslint-react": "^3.5.1",
+      "typescript": "^3.0.1"
+    },
+    "peerDependencies": {
+      "react": "^16.6.0"  
+    }
+  }
+
