+++ packages/fast-data-utilities-react/src/__tests__/schemas/any-of.schema.json
@@ -1,0 +1,93 @@
+{
+    "$schema": "http://json-schema.org/schema#",
+    "title": "Component with anyOf",
+    "description": "A test component's schema definition.",
+    "type": "object",
+    "id": "anyOf",
+    "anyOf": [
+        {
+            "description": "String",
+            "type": "object",
+            "properties": {
+                "string": {
+                    "title": "String",
+                    "type": "string"
+                }
+            },
+            "required": [
+                "string"
+            ]
+        },
+        {
+            "description": "Number",
+            "type": "object",
+            "properties": {
+                "number": {
+                    "title": "Number",
+                    "type": "number"
+                }
+            },
+            "required": [
+                "number"
+            ]
+        },
+        {
+            "type": "object",
+            "properties": {
+                "nestedAnyOf": {
+                    "title": "Nested anyOf",
+                    "anyOf": [
+                        {
+                            "description": "Object",
+                            "type": "object",
+                            "properties": {
+                                "object": {
+                                    "title": "String",
+                                    "type": "object",
+                                    "properties": {
+                                        "string": {
+                                            "title": "String",
+                                            "type": "string"
+                                        }
+                                    },
+                                    "required": [
+                                        "string"
+                                    ]
+                                }
+                            },
+                            "required": [
+                                "object"
+                            ]
+                        },
+                        {
+                            "description": "String",
+                            "type": "object",
+                            "properties": {
+                                "string": {
+                                    "title": "String",
+                                    "type": "string"
+                                }
+                            },
+                            "required": [
+                                "string"
+                            ]
+                        },
+                        {
+                            "description": "Number",
+                            "type": "object",
+                            "properties": {
+                                "number": {
+                                    "title": "Number",
+                                    "type": "number"
+                                }
+                            },
+                            "required": [
+                                "number"
+                            ]
+                        }
+                    ]
+                }
+            }
+        }
+    ]
+}
