+++ packages/fast-data-utilities-react/src/__tests__/schemas/arrays.schema.json
@@ -1,0 +1,39 @@
+{
+    "$schema": "http://json-schema.org/schema#",
+    "title": "Component with array",
+    "description": "A test component's schema definition.",
+    "type": "object",
+    "id": "arrays",
+    "properties": {
+        "strings": {
+            "title": "Array of strings",
+            "type": "array",
+            "items": {
+                "title": "String",
+                "type": "string"
+            },
+            "minItems": 2,
+            "maxItems": 5
+        },
+        "objects": {
+            "title": "Array of objects",
+            "type": "array",
+            "items": {
+                "title": "Object",
+                "type": "object",
+                "properties": {
+                    "string": {
+                        "title": "String",
+                        "type": "string"
+                    }
+                },
+                "required": [
+                    "string"
+                ]
+            }
+        }
+    },
+    "required": [
+        "strings"
+    ]
+}
