+++ packages/fast-form-generator-react/package.json
@@ -39,7 +39,7 @@
       "global": {
         "statements": 59,
         "branches": 37,
-        "functions": 53,
+        "functions": 50,
         "lines": 59
       }
     },
@@ -113,5 +113,8 @@
     "webpack-cli": "^3.1.1",
     "webpack-dev-server": "^3.1.3",
     "yargs": "^11.0.0"
+  },
+  "dependencies": {
+    "@microsoft/fast-data-utilities-react": "^1.0.0"
   }
 }
