+++ packages/fast-form-generator-react/src/form/index.ts
@@ -1,5 +1,5 @@
 import Form from "./form";
-import { mapDataToComponent } from "./form.utilities";
+import { mapDataToComponent } from "@microsoft/fast-data-utilities-react";
 
 export default Form;
 export { mapDataToComponent };
